Crate dslab_network

A library for modeling network communications. It can be easily plugged in simulations based on DSLab core by defining network nodes and binding the simulation components to the nodes. After that, using the library one can simulate data transfers or messaging between the components. The library includes several ready-to-use network models and allows to use it with custom model implementations.

Network models

ConstantBandwidthNetworkModel: Simple topology-unaware model where each transfer gets the full network bandwidth, i.e. there is no contention.

SharedBandwidthNetworkModel: Topology-unaware model where the network bandwidth is shared fairly among all current transfers.

TopologyAwareNetworkModel: Topology-aware model which uses information about the network Topology (links connecting the nodes) and relies on RoutingAlgorithm to compute paths between the nodes. The link’s bandwidth is shared fairly among the transfers using the link.

Examples

network-simple: demonstrates the use of topology-unaware models.

network-topology: demonstrates the use of TopologyAwareNetworkModel.

network-topology-benchmark: benchmarks the performance of TopologyAwareNetworkModel on different topologies.

Re-exports

pub use link::BandwidthSharingPolicy;

pub use link::Link;

pub use link::LinkId;

pub use model::DataTransfer;

pub use model::DataTransferCompleted;

pub use model::NetworkModel;

pub use network::Message;

pub use network::MessageDelivered;

pub use network::Network;

pub use node::Node;

pub use node::NodeId;

pub use topology::Topology;

Modules

link

Network link.

model

Network model interface.

models

Network model implementations.

network

Simulation component representing a network.

node

Network node.

routing

Routing algorithms.

topology

Network topology.