Module dslab_network::models

Network model implementations.

Re-exports

pub use constant::ConstantBandwidthNetworkModel;

pub use shared::SharedBandwidthNetworkModel;

pub use topology_aware::TopologyAwareNetworkModel;

Modules

constant

Network model without congestion where each transfer gets the full bandwidth.

shared

Network model where the bandwidth is shared fairly among all current transfers.

topology_aware

Topology-aware network model.