Struct dslab_network::network::Network

pub struct Network { /* private fields */ }

Simulation component representing a network.

This is the main entry point for all network operations, which relies internally on the supplied network model.

Implementations

impl Network

pub fn new(model: Box<dyn NetworkModel>, ctx: SimulationContext) -> Self

Creates a new network with supplied network model.

pub fn add_node<S>( &mut self, name: S, local_model: Box<dyn NetworkModel> ) -> NodeId
where S: Into<String>,

Adds a new network node.

The supplied local_model is used to model the intra-node communications.

pub fn get_node_id(&self, node: &str) -> usize

Returns the node id by its name.

pub fn get_nodes(&self) -> Vec<String>

Returns the list of network nodes.

pub fn add_link(&mut self, node1: &str, node2: &str, link: Link) -> LinkId

Adds a new bidirectional link between two nodes.

pub fn add_unidirectional_link( &mut self, node_from: &str, node_to: &str, link: Link ) -> LinkId

Adds a new unidirectional link between two nodes.

pub fn add_full_duplex_link( &mut self, node1: &str, node2: &str, link: Link ) -> (LinkId, LinkId)

Adds two unidirectional links with the same parameters between two nodes in opposite directions.

This allows to model the full-duplex communication links.

pub fn init_topology(&mut self)

Performs initialization of network topology, such as computing the paths between the nodes.

Must be called after all links are added and before submitting any operations.

pub fn set_location(&mut self, id: Id, node: &str)

Sets the location of the simulation component id to the node node.

pub fn get_location_opt(&self, id: Id) -> Option<NodeId>

Returns the location (node id) of the simulation component if it is set.

pub fn get_location(&self, id: Id) -> NodeId

Returns the location (node id) of the simulation component.

Panics if the component location is not set.

pub fn bandwidth(&self, src: Id, dst: Id) -> f64

Returns the network bandwidth between two simulation components.

pub fn latency(&self, src: Id, dst: Id) -> f64

Returns the network latency between two simulation components.

pub fn transfer_data( &mut self, src: Id, dst: Id, size: f64, notification_dst: Id ) -> usize

Starts a data transfer between two simulation components, returns unique transfer id.

The network locations of these components must be previously registered via Self::set_location. The transfer completion time is calculated by the underlying network model. The DataTransferCompleted event is sent to notification_dst on the transfer completion.

pub fn send_msg(&mut self, message: String, src: Id, dst: Id) -> usize

Sends a message between two simulation components, returns unique message id.

The network locations of these components must be previously registered via Self::set_location. The message delivery time is equal to the network latency, assuming the message data has a small size. The MessageDelivered event is sent to dst on the message delivery.

pub fn send_event<T: EventData>(&mut self, data: T, src: Id, dst: Id) -> EventId

Sends an event between two simulation components, returns unique event id.

The network locations of these components must be previously registered via Self::set_location. The event delivery time is equal to the network latency, assuming the event data has a small size.