Trait dslab_network::model::NetworkModel

pub trait NetworkModel {
    // Required methods
    fn is_topology_aware(&self) -> bool;
    fn bandwidth(&self, src: NodeId, dst: NodeId) -> f64;
    fn latency(&self, src: NodeId, dst: NodeId) -> f64;
    fn start_transfer(&mut self, dt: DataTransfer, ctx: &mut SimulationContext);
    fn on_transfer_completion(
        &mut self,
        dt: DataTransfer,
        ctx: &mut SimulationContext
    );

    // Provided methods
    fn topology(&self) -> Option<&Topology> { ... }
    fn topology_mut(&mut self) -> Option<&mut Topology> { ... }
    fn on_topology_change(&mut self, _ctx: &mut SimulationContext) { ... }
}

Network model interface.

The main functions of the network model:

Provide bandwidth and latency between network nodes,

Calculate data transfer completion times and emit DataTransferCompleted events.

A topology-aware model uses information about the network topology (links connecting the nodes) and relies on a routing algorithm to compute paths between the nodes.

Required Methods

fn is_topology_aware(&self) -> bool

Returns true is the model is topology-aware.

fn bandwidth(&self, src: NodeId, dst: NodeId) -> f64

Returns the network bandwidth from node src to node dst.

fn latency(&self, src: NodeId, dst: NodeId) -> f64

Returns the network latency from node src to node dst.

fn start_transfer(&mut self, dt: DataTransfer, ctx: &mut SimulationContext)

Starts data transfer.

Must calculate the transfer completion time and emit the DataTransferCompleted event at this time. The event must be emitted via the passed simulation context using [SimulationContext::emit_self].

fn on_transfer_completion( &mut self, dt: DataTransfer, ctx: &mut SimulationContext )

Callback for notifying the model about data transfer completion.

This is necessary since the model itself does not receive the DataTransferCompleted event.

Provided Methods

fn topology(&self) -> Option<&Topology>

Returns a reference to inner network topology.

Must be implemented for topology-aware model.

fn topology_mut(&mut self) -> Option<&mut Topology>

Returns a mutable reference to inner network topology.

Must be implemented for topology-aware model.

fn on_topology_change(&mut self, _ctx: &mut SimulationContext)

Callback for notifying topology-aware model about the topology change.

Must be implemented for topology-aware model.